refactor: errorLogWriter 리팩토링
