--- src/backend/src/main/java/com/graduate/odondong/config/FilterChainProxyAdvice.java
@@ -1,56 +1,0 @@
-package com.graduate.odondong.config;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.annotation.Around;
-import org.aspectj.lang.annotation.Aspect;
-import org.springframework.security.web.firewall.RequestRejectedException;
-import org.springframework.stereotype.Component;
-
-import lombok.extern.slf4j.Slf4j;
-
-import static com.graduate.odondong.util.ErrorLogWriter.writeExceptionWithRequest;
-
-@Aspect
-@Component
-@Slf4j
-public class FilterChainProxyAdvice {
-
-	@Around("execution(public void org.springframework.security.web.FilterChainProxy.doFilter(..))")
-	public void handleRequestRejectedException(ProceedingJoinPoint pjp) throws Throwable {
-		try {
-			pjp.proceed();
-		} catch (RequestRejectedException e) {
-			HttpServletRequest request = (HttpServletRequest)pjp.getArgs()[0];
-			String ip = request.getHeader("X-Forwarded-For");
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("Proxy-Client-IP");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("WL-Proxy-Client-IP");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("HTTP_CLIENT_IP");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("HTTP_X_FORWARDED_FOR");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("X-Real-IP");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("X-RealIP");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getHeader("REMOTE_ADDR");
-			}
-			if (ip == null || ip.length() == 0 || "unknown".equalsIgnoreCase(ip)) {
-				ip = request.getRemoteAddr();
-			}
-			writeExceptionWithRequest(e, request);
-			log.error("ip 주소는 " + ip + "입니다");
-		}
-	}
-}
-
